It's possible to add enough occurrences to a postings list that it overflows the integer count of total occurrences.  So, we make the count a long and peg it to Integer.MAX_VALUE when necessary.  Also, go ahead and make sure that we throw an arithmetic exception when attempting to encode negative numbers when only positive numbers are valid.
Fixes for issue 4: https://minion.dev.java.net/issues/show_bug.cgi?id=4

+++ src/com/sun/labs/minion/util/buffer/NIOBuffer.java
@@ -24,13 +24,13 @@
 
 package com.sun.labs.minion.util.buffer;
 
+import com.sun.labs.minion.util.ChannelUtil;
 import java.io.DataOutput;
 import java.io.OutputStream;
 
 import java.nio.ByteBuffer;
 import java.nio.channels.WritableByteChannel;
 
-import com.sun.labs.minion.util.*;
 
 /**
  * A buffer that uses java.nio buffers as the backing store.
@@ -398,7 +398,7 @@ public class NIOBuffer extends StdBufferImpl {
 
         while((curr & 0x80) != 0) {
             curr = units.get();
-            res |= ((int) (curr  & 0x7F)) << shift;
+            res |= ((curr & 0x7F)) << shift;
             shift += 7;
         }
         return res;
